refactor: generalizes data formats
